# docs/api.yaml
openapi: 3.0.1
info:
  title: Achievements API
  description: This is a an API for managing Achievements for games
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - description: Localhost server for testing API
    url: http://localhost:8080
...
paths:
...
  /games/{gameID}:
    description: The resource that contains a specific game
    parameters:
      - name: gameID
        in: path
        required: true
        description: The ID of the game to retrieve.
        schema:
          type: integer
    get:
      tags:
        - Games
      summary: Get a specific game by ID
      description: Returns the details of a specific game.
      responses:
        200:
          description: In case of success, returns the details of the game.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        404:
          $ref: "#/components/responses/404NotFound"
    put:
      tags:
        - Games
      summary: Update a game
      description: Update the details of a specific game.
      operationId: updateGame
      parameters:
        - name: gameId
          in: path
          description: Id of the group to be updated
          required: true
          schema:
            type: integer
      requestBody:
        description: Game data to update
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewGame"
    delete:
      tags:
        - Games
      summary: Delete a specific game by ID
      description: Deletes a specific game by ID.
      responses:
        204:
          description: Game deleted successfully.
        404:
          $ref: "#/components/responses/404NotFound"
...